+++ pkg/config/config_test.go
@@ -3,6 +3,7 @@ package config
 import (
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
+	"os"
 	"testing"
 )
 
@@ -11,6 +12,9 @@ func TestNewOperatorConfig(t *testing.T) {
 		// given
 		version := "0.0.0"
 		t.Setenv("NAMESPACE", "ecosystem")
+		t.Setenv("ARCHIVE_VOLUME_DOWNLOAD_SERVICE_NAME", "service")
+		t.Setenv("ARCHIVE_VOLUME_DOWNLOAD_SERVICE_PROTOCOL", "http")
+		t.Setenv("ARCHIVE_VOLUME_DOWNLOAD_SERVICE_PORT", "8080")
 
 		// when
 		operatorConfig, err := NewOperatorConfig(version)
@@ -19,11 +23,14 @@ func TestNewOperatorConfig(t *testing.T) {
 		require.NoError(t, err)
 		require.NotNil(t, operatorConfig)
 	})
-	t.Run("should succeed with namespace set", func(t *testing.T) {
+	t.Run("should succeed with stage set", func(t *testing.T) {
 		// given
 		version := "0.0.0"
 		t.Setenv("NAMESPACE", "ecosystem")
 		t.Setenv("STAGE", "development")
+		t.Setenv("ARCHIVE_VOLUME_DOWNLOAD_SERVICE_NAME", "service")
+		t.Setenv("ARCHIVE_VOLUME_DOWNLOAD_SERVICE_PROTOCOL", "http")
+		t.Setenv("ARCHIVE_VOLUME_DOWNLOAD_SERVICE_PORT", "8080")
 
 		// when
 		operatorConfig, err := NewOperatorConfig(version)
@@ -45,9 +52,10 @@ func TestNewOperatorConfig(t *testing.T) {
 		require.Nil(t, operatorConfig)
 		assert.ErrorContains(t, err, "failed to parse version: Invalid Semantic Version")
 	})
-	t.Run("fail to read namespace", func(t *testing.T) {
+	t.Run("fail to read namespace because of non existent env var", func(t *testing.T) {
 		// given
 		version := "0.0.0"
+		require.NoError(t, os.Unsetenv("NAMESPACE"))
 
 		// when
 		operatorConfig, err := NewOperatorConfig(version)