+++ pkg/reconciler/support_archive_controller.go
@@ -4,24 +4,32 @@ import (
 	"context"
 	"errors"
 	"fmt"
-	k8sv1 "github.com/cloudogu/k8s-support-archive-lib/api/v1"
-	"k8s.io/apimachinery/pkg/runtime"
+	libv1 "github.com/cloudogu/k8s-support-archive-lib/api/v1"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	ctrl "sigs.k8s.io/controller-runtime"
+	"sigs.k8s.io/controller-runtime/pkg/client"
 	"sigs.k8s.io/controller-runtime/pkg/controller"
 	"sigs.k8s.io/controller-runtime/pkg/log"
 	"sigs.k8s.io/controller-runtime/pkg/predicate"
 	"sigs.k8s.io/controller-runtime/pkg/reconcile"
+	"slices"
+)
+
+const (
+	finalizerName = "k8s.cloudogu.com/support-archive-reconciler"
 )
 
 type SupportArchiveReconciler struct {
-	Client EcosystemClientSet
-	Scheme *runtime.Scheme
+	client         supportArchiveV1Interface
+	archiveHandler archiveHandler
+	cleaner        archiveCleaner
 }
 
-func NewSupportArchiveReconciler(client EcosystemClientSet, scheme *runtime.Scheme) *SupportArchiveReconciler {
+func NewSupportArchiveReconciler(client supportArchiveV1Interface, archiveHandler archiveHandler, cleaner archiveCleaner) *SupportArchiveReconciler {
 	return &SupportArchiveReconciler{
-		Client: client,
-		Scheme: scheme,
+		client:         client,
+		archiveHandler: archiveHandler,
+		cleaner:        cleaner,
 	}
 }
 
@@ -30,7 +38,37 @@ func (s *SupportArchiveReconciler) Reconcile(ctx context.Context, req ctrl.Reque
 
 	logger.Info(fmt.Sprintf("Reconciler is triggered by resource %q", req.NamespacedName))
 
-	return ctrl.Result{}, nil
+	archiveInterface := s.client.SupportArchives(req.Namespace)
+	cr, err := archiveInterface.Get(ctx, req.Name, metav1.GetOptions{})
+	if err != nil {
+		return ctrl.Result{}, client.IgnoreNotFound(err)
+	}
+
+	if !slices.Contains(cr.GetFinalizers(), finalizerName) {
+		logger.Info(fmt.Sprintf("Adding finalizer to support archive %q", cr.Name))
+		cr.Finalizers = append(cr.Finalizers, finalizerName)
+		_, updateErr := archiveInterface.Update(ctx, cr, metav1.UpdateOptions{})
+		return ctrl.Result{Requeue: true}, updateErr
+	}
+
+	if !cr.GetDeletionTimestamp().IsZero() {
+		cleanupErr := s.cleaner.Clean(ctx, cr.Name, cr.Namespace)
+		if cleanupErr != nil {
+			// Do not return here to avoid blocking in error case.
+			// Garbage collection can try to clean up inconsistent files later.
+			logger.Info(fmt.Sprintf("Failed to clean up for support archive request %q: %v", cr.Name, cleanupErr))
+		}
+
+		_, updateErr := archiveInterface.RemoveFinalizer(ctx, cr, finalizerName)
+		if updateErr != nil {
+			return ctrl.Result{Requeue: true}, updateErr
+		}
+
+		return ctrl.Result{Requeue: false}, nil
+	}
+
+	requeue, err := s.archiveHandler.HandleArchiveRequest(ctx, cr)
+	return ctrl.Result{Requeue: requeue}, err
 }
 
 // SetupWithManager sets up the controller with the Manager.
@@ -48,6 +86,6 @@ func (s *SupportArchiveReconciler) SetupWithManager(mgr ctrl.Manager) error {
 	return ctrl.NewControllerManagedBy(mgr).
 		WithEventFilter(predicate.GenerationChangedPredicate{}).
 		WithOptions(options).
-		For(&k8sv1.SupportArchive{}).
+		For(&libv1.SupportArchive{}).
 		Complete(s)
 }