+++ pkg/reconciler/support_archive_controller_test.go
@@ -6,6 +6,8 @@ import (
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/mock"
 	"github.com/stretchr/testify/require"
+	"k8s.io/apimachinery/pkg/api/errors"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/runtime"
 	"k8s.io/apimachinery/pkg/runtime/schema"
 	"k8s.io/apimachinery/pkg/types"
@@ -14,11 +16,15 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/config"
 	"sigs.k8s.io/controller-runtime/pkg/log"
 	"testing"
+	"time"
 )
 
 var testCtx = context.Background()
 
-const testSupportArchive = "test-support-archive"
+const (
+	testSupportArchive = "test-support-archive"
+	testNamespace      = "test-namespace"
+)
 
 func TestNewSupportArchiveOperator(t *testing.T) {
 	// override default controller method to retrieve a kube config
@@ -28,10 +34,10 @@ func TestNewSupportArchiveOperator(t *testing.T) {
 
 	t.Run("success", func(t *testing.T) {
 		// given
-		mockClient := NewMockEcosystemClientSet(t)
+		mockClient := newMockSupportArchiveV1Interface(t)
 
 		// when
-		doguManager := NewSupportArchiveReconciler(mockClient, getTestScheme())
+		doguManager := NewSupportArchiveReconciler(mockClient, newMockArchiveHandler(t), newMockArchiveCleaner(t))
 
 		// then
 		require.NotNil(t, doguManager)
@@ -40,12 +46,6 @@ func TestNewSupportArchiveOperator(t *testing.T) {
 
 func createTestRestConfig() (*rest.Config, error) {
 	return &rest.Config{}, nil
-}
-
-func getTestScheme() *runtime.Scheme {
-	scheme := runtime.NewScheme()
-
-	return scheme
 }
 
 func TestSupportArchiveReconciler_SetupWithManager(t *testing.T) {
@@ -92,16 +92,192 @@ func createScheme(t *testing.T) *runtime.Scheme {
 }
 
 func TestSupportArchiveReconciler_Reconcile(t *testing.T) {
-	t.Run("should succeed", func(t *testing.T) {
+	initialArchiveCr := &v1.SupportArchive{
+		ObjectMeta: metav1.ObjectMeta{
+			Name:      testSupportArchive,
+			Namespace: testNamespace,
+		},
+	}
+
+	archiveCrWithFinalizer := &v1.SupportArchive{
+		ObjectMeta: metav1.ObjectMeta{
+			Name:       testSupportArchive,
+			Namespace:  testNamespace,
+			Finalizers: []string{"k8s.cloudogu.com/support-archive-reconciler"},
+		},
+	}
+
+	deletedArchiveCr := &v1.SupportArchive{
+		ObjectMeta: metav1.ObjectMeta{
+			Name:              testSupportArchive,
+			Namespace:         testNamespace,
+			Finalizers:        []string{"k8s.cloudogu.com/support-archive-reconciler"},
+			DeletionTimestamp: &metav1.Time{Time: time.Now()},
+		},
+	}
+
+	t.Run("should do nothing if the request is not found", func(t *testing.T) {
 		// given
-		request := ctrl.Request{NamespacedName: types.NamespacedName{Name: testSupportArchive}}
-		sut := &SupportArchiveReconciler{}
+		request := ctrl.Request{NamespacedName: types.NamespacedName{Name: testSupportArchive, Namespace: testNamespace}}
+		mockV1Interface := newMockSupportArchiveV1Interface(t)
+		mockInterface := newMockSupportArchiveInterface(t)
+		mockV1Interface.EXPECT().SupportArchives(testNamespace).Return(mockInterface)
+		mockInterface.EXPECT().Get(testCtx, testSupportArchive, metav1.GetOptions{}).Return(nil, errors.NewNotFound(schema.GroupResource{}, "not found"))
+
+		sut := &SupportArchiveReconciler{
+			client: mockV1Interface,
+		}
 
 		// when
 		actual, err := sut.Reconcile(testCtx, request)
 
 		// then
 		require.NoError(t, err)
-		assert.Equal(t, ctrl.Result{}, actual)
+		assert.Equal(t, ctrl.Result{Requeue: false}, actual)
+	})
+
+	t.Run("should add finalizer and retry", func(t *testing.T) {
+		// given
+		request := ctrl.Request{NamespacedName: types.NamespacedName{Name: testSupportArchive, Namespace: testNamespace}}
+		mockV1Interface := newMockSupportArchiveV1Interface(t)
+		mockInterface := newMockSupportArchiveInterface(t)
+		mockInterface.EXPECT().Get(testCtx, testSupportArchive, metav1.GetOptions{}).Return(initialArchiveCr, nil)
+		mockInterface.EXPECT().Update(testCtx, archiveCrWithFinalizer, metav1.UpdateOptions{}).Return(nil, nil)
+
+		mockV1Interface.EXPECT().SupportArchives(testNamespace).Return(mockInterface)
+		sut := &SupportArchiveReconciler{
+			client: mockV1Interface,
+		}
+
+		// when
+		actual, err := sut.Reconcile(testCtx, request)
+
+		// then
+		require.NoError(t, err)
+		assert.Equal(t, ctrl.Result{Requeue: true}, actual)
+	})
+
+	t.Run("should retry with error on error adding finalizer", func(t *testing.T) {
+		// given
+		archiveCrWithoutFinalizer := &v1.SupportArchive{
+			ObjectMeta: metav1.ObjectMeta{
+				Name:      testSupportArchive,
+				Namespace: testNamespace,
+			},
+		}
+
+		request := ctrl.Request{NamespacedName: types.NamespacedName{Name: testSupportArchive, Namespace: testNamespace}}
+		mockV1Interface := newMockSupportArchiveV1Interface(t)
+		mockInterface := newMockSupportArchiveInterface(t)
+		mockInterface.EXPECT().Get(testCtx, testSupportArchive, metav1.GetOptions{}).Return(archiveCrWithoutFinalizer, nil)
+		mockInterface.EXPECT().Update(testCtx, archiveCrWithFinalizer, metav1.UpdateOptions{}).Return(nil, assert.AnError)
+
+		mockV1Interface.EXPECT().SupportArchives(testNamespace).Return(mockInterface)
+		sut := &SupportArchiveReconciler{
+			client: mockV1Interface,
+		}
+
+		// when
+		actual, err := sut.Reconcile(testCtx, request)
+
+		// then
+		require.Error(t, err)
+		assert.ErrorIs(t, err, assert.AnError)
+		assert.Equal(t, ctrl.Result{Requeue: true}, actual)
+	})
+
+	t.Run("should proceed with archive creation", func(t *testing.T) {
+		// given
+		request := ctrl.Request{NamespacedName: types.NamespacedName{Name: testSupportArchive, Namespace: testNamespace}}
+		mockV1Interface := newMockSupportArchiveV1Interface(t)
+		mockInterface := newMockSupportArchiveInterface(t)
+		mockV1Interface.EXPECT().SupportArchives(testNamespace).Return(mockInterface)
+		mockInterface.EXPECT().Get(testCtx, testSupportArchive, metav1.GetOptions{}).Return(archiveCrWithFinalizer, nil)
+		archiveHandlerMock := newMockArchiveHandler(t)
+		archiveHandlerMock.EXPECT().HandleArchiveRequest(testCtx, archiveCrWithFinalizer).Return(true, nil)
+
+		sut := &SupportArchiveReconciler{
+			client:         mockV1Interface,
+			archiveHandler: archiveHandlerMock,
+		}
+
+		// when
+		actual, err := sut.Reconcile(testCtx, request)
+
+		// then
+		require.NoError(t, err)
+		assert.Equal(t, ctrl.Result{Requeue: true}, actual)
+	})
+
+	t.Run("should not requeue if archive handler completed archive creation", func(t *testing.T) {
+		// given
+		request := ctrl.Request{NamespacedName: types.NamespacedName{Name: testSupportArchive, Namespace: testNamespace}}
+		mockV1Interface := newMockSupportArchiveV1Interface(t)
+		mockInterface := newMockSupportArchiveInterface(t)
+		mockV1Interface.EXPECT().SupportArchives(testNamespace).Return(mockInterface)
+		mockInterface.EXPECT().Get(testCtx, testSupportArchive, metav1.GetOptions{}).Return(archiveCrWithFinalizer, nil)
+		archiveHandlerMock := newMockArchiveHandler(t)
+		archiveHandlerMock.EXPECT().HandleArchiveRequest(testCtx, archiveCrWithFinalizer).Return(false, nil)
+
+		sut := &SupportArchiveReconciler{
+			client:         mockV1Interface,
+			archiveHandler: archiveHandlerMock,
+		}
+
+		// when
+		actual, err := sut.Reconcile(testCtx, request)
+
+		// then
+		require.NoError(t, err)
+		assert.Equal(t, ctrl.Result{Requeue: false}, actual)
+	})
+
+	t.Run("should cleanup if deletion timestamp is set", func(t *testing.T) {
+		// given
+		request := ctrl.Request{NamespacedName: types.NamespacedName{Name: testSupportArchive, Namespace: testNamespace}}
+		mockV1Interface := newMockSupportArchiveV1Interface(t)
+		mockInterface := newMockSupportArchiveInterface(t)
+		mockV1Interface.EXPECT().SupportArchives(testNamespace).Return(mockInterface)
+		mockInterface.EXPECT().Get(testCtx, testSupportArchive, metav1.GetOptions{}).Return(deletedArchiveCr, nil)
+		archiveCleanerMock := newMockArchiveCleaner(t)
+		archiveCleanerMock.EXPECT().Clean(testCtx, testSupportArchive, testNamespace).Return(nil)
+		mockInterface.EXPECT().RemoveFinalizer(testCtx, deletedArchiveCr, "k8s.cloudogu.com/support-archive-reconciler").Return(nil, nil)
+
+		sut := &SupportArchiveReconciler{
+			client:  mockV1Interface,
+			cleaner: archiveCleanerMock,
+		}
+
+		// when
+		actual, err := sut.Reconcile(testCtx, request)
+
+		// then
+		require.NoError(t, err)
+		assert.Equal(t, ctrl.Result{Requeue: false}, actual)
+	})
+
+	t.Run("should requeue and not block on cleanup errors", func(t *testing.T) {
+		// given
+		request := ctrl.Request{NamespacedName: types.NamespacedName{Name: testSupportArchive, Namespace: testNamespace}}
+		mockV1Interface := newMockSupportArchiveV1Interface(t)
+		mockInterface := newMockSupportArchiveInterface(t)
+		mockV1Interface.EXPECT().SupportArchives(testNamespace).Return(mockInterface)
+		mockInterface.EXPECT().Get(testCtx, testSupportArchive, metav1.GetOptions{}).Return(deletedArchiveCr, nil)
+		archiveCleanerMock := newMockArchiveCleaner(t)
+		archiveCleanerMock.EXPECT().Clean(testCtx, testSupportArchive, testNamespace).Return(assert.AnError)
+		mockInterface.EXPECT().RemoveFinalizer(testCtx, deletedArchiveCr, "k8s.cloudogu.com/support-archive-reconciler").Return(nil, assert.AnError)
+
+		sut := &SupportArchiveReconciler{
+			client:  mockV1Interface,
+			cleaner: archiveCleanerMock,
+		}
+
+		// when
+		actual, err := sut.Reconcile(testCtx, request)
+
+		// then
+		require.Error(t, err)
+		assert.ErrorIs(t, err, assert.AnError)
+		assert.Equal(t, ctrl.Result{Requeue: true}, actual)
 	})
 }